add check before moving if mob has equipped thier armor
encountered behavior that when including _ready(), godot seems to process the statemachine transitions before the enemy scene can get the params export.

this is a bandaid fix but it works. when the enemy is being instantiated by a spawner we can change it so that equip armor is called shortly after instantiation